Applies a hotfix for the issue causing the app not being properly closed.
When the app is started minimzed, brought to the foreground and we then request app termination via the [Quit] action in the context menu of the notification area icon the app window(s) would terminate, cleanup would be performed but the app process would still be running (visible in the TaskManager). This commit adds a hotfix for this issue.